Moved HomeNotificationViewModel+Network for both News and Message into one file and renamed into HomeNotificationViewModel+Network
