+++ Laundron-User/LUFeature/LUHome/LUHome/LUHome/HomeNotification/HomeNotificationViewModel+Network.swift
@@ -1,5 +1,5 @@
 //
-//  NotificationMessageViewModel+Network.swift
+//  NotificationNewsViewModel+Network.swift
 //  MainApp
 //
 //  Created by Evan Christian on 10/10/20.
@@ -7,9 +7,33 @@
 //
 
 import Foundation
-import LUHomeNetwork
-import LUCodable
 import LUHandler
+import LUCodable
+import LUHomeNetwork
+
+extension HomeNotificationViewModel: RetrieveNotificationNewsDelegate {
+    func didSuccessRetrieveNotificationNews(response: HomeResponse.GetNotificationNewsResponse) {
+        var data = [HomeNotificationNewsData]()
+        if response.newsList.count > 1 {
+            for news in response.newsList {
+                guard let newsContent = news.message else { return }
+                data.append(HomeNotificationNewsData.init(message: newsContent))
+            }
+            notificationNewsDataSource?.setData(datas: data)
+            view?.reloadNewsTableData()
+        } else {
+            view?.newsTableContentIsEmpty = true
+        }
+        view?.stopLoading()
+    }
+    
+    func didFailedRetrieveNotificationNews(error: Error) {
+        view?.newsTableContentIsEmpty = true
+        view?.newsTableIsEmpty()
+        view?.stopLoading()
+        Log.info(message: error)
+    }
+}
 
 extension HomeNotificationViewModel: RetrieveNotificationMessageDelegate {
     func didSuccessRetrieveNotificationMessage(response: HomeResponse.GetNotificationMessageResponse) {
